# openapi.json
{"openapi":"3.0.0","paths":{"/auth/signup":{"post":{"operationId":"signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"responses":{"201":{"description":""}}}},"/auth/signin":{"post":{"operationId":"signin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDtoSignin"}}}},"responses":{"201":{"description":""}}}},"/auth/updatePassword":{"post":{"operationId":"updatePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDtoUpdatePassword"}}}},"responses":{"201":{"description":""}}}},"/auth/signout":{"get":{"operationId":"signout","parameters":[],"responses":{"200":{"description":""}}}},"/users/me/{id}":{"get":{"operationId":"getMyUser","parameters":[],"responses":{"200":{"description":""}}}},"/users":{"get":{"operationId":"getUsers","parameters":[],"responses":{"200":{"description":""}}}},"/tasks/{authorId}":{"get":{"operationId":"getMyTask","parameters":[],"responses":{"200":{"description":""}}}},"/tasks/{authorId}/{filter}":{"get":{"operationId":"getFilterTask","parameters":[],"responses":{"200":{"description":""}}}},"/tasks/{id}/create":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskDto"}}}},"responses":{"201":{"description":""}}}},"/tasks/{id}/update/{taskId}":{"post":{"operationId":"updateTask","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskUpdate"}}}},"responses":{"201":{"description":""}}}},"/tasks/{id}/delete/{taskId}":{"get":{"operationId":"deleteTask","parameters":[],"responses":{"200":{"description":""}}}},"/tasks":{"get":{"operationId":"getTasks","parameters":[],"responses":{"200":{"description":""}}}}},"info":{"title":"My API","description":"API","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"AuthDto":{"type":"object","properties":{}},"AuthDtoSignin":{"type":"object","properties":{}},"AuthDtoUpdatePassword":{"type":"object","properties":{}},"TaskDto":{"type":"object","properties":{}},"TaskUpdate":{"type":"object","properties":{}}}}}